Change variable int to long in interface 'Add' and 'Subtract', updated implection in class 'Math'

+++ src/com/thecalculator/Add.java
@@ -2,5 +2,5 @@ package com.thecalculator;
 
 public interface Add {
 
-    int add(int number);
+    long add(long number);
 }
